Convert np.int* to np.int, because it doesn't benchmark any differently

+++ crawlai/model/environment.py
@@ -16,7 +16,7 @@ class CritterEnvironment(PyEnvironment):
 
 		# TODO: Figure out how to get len(AiCritterMixin.CHOICES)
 		self._action_spec = array_spec.BoundedArraySpec(
-			shape=(), dtype=np.int32,
+			shape=(), dtype=np.int,
 			minimum=0, maximum=n_choices - 1,
 			name="action")
